attempt by swapping to str_to_bool usage
